fix: boss hit flash renders red, not magenta
The flash tint built `new XnaColor(int, int, int, float)` (Alpha is float),
which bound the (float,float,float,float) 0-1 overload and clamped any
channel >1 to 255. White flashes (all channels max) looked fine, but the
red boss flash Color8 (255,0,77) became (255,0,255) = magenta. Cast Alpha
to int so it binds the 0-255 overload; boss flashes red, white flashes
render as pico-8 white (Color7).

+++ Systems/FlashRenderingSystem.cs
@@ -65,7 +65,15 @@ namespace CherryBomb.Systems
 				}
 
 				var mask = GetMask(sprite.CurrentFrame);
-				var tint = new XnaColor(flash.Color.R, flash.Color.G, flash.Color.B, flash.Alpha);
+				// Cast Alpha to int so this binds Color(int,int,int,int) (0-255), not
+				// the (float,float,float,float) 0-1 overload — which clamped any channel
+				// >1 to 255 and turned the red boss flash (255,0,77) into magenta.
+				var tint = new XnaColor(
+					flash.Color.R,
+					flash.Color.G,
+					flash.Color.B,
+					(int)flash.Alpha
+				);
 
 				_spriteBatch.Draw(
 					mask,
